+++ public/js/resources_js_pages_TestStyle_js.js
@@ -1,0 +1,108 @@
+"use strict";
+(self["webpackChunk"] = self["webpackChunk"] || []).push([["resources_js_pages_TestStyle_js"],{
+
+/***/ "./resources/js/components/styles/Button.styled.js":
+/*!*********************************************************!*\
+  !*** ./resources/js/components/styles/Button.styled.js ***!
+  \*********************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+__webpack_require__.r(__webpack_exports__);
+/* harmony export */ __webpack_require__.d(__webpack_exports__, {
+/* harmony export */   "Button": () => (/* binding */ Button),
+/* harmony export */   "ButtonBackUp": () => (/* binding */ ButtonBackUp)
+/* harmony export */ });
+/* harmony import */ var styled_components__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! styled-components */ "./node_modules/styled-components/dist/styled-components.browser.esm.js");
+var _templateObject, _templateObject2;
+
+function _taggedTemplateLiteral(strings, raw) { if (!raw) { raw = strings.slice(0); } return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }
+
+
+var Button = styled_components__WEBPACK_IMPORTED_MODULE_0__["default"].button(_templateObject || (_templateObject = _taggedTemplateLiteral(["\n  border-radius: 2px;\n  border: none;\n  cursor: pointer;\n  font-weight: 700;\n  height: 40px;\n  width: 100%;\n  background-color: ", ";\n  color: ", ";\n  font-family: 'Gotham Black';\n  max-width: 350px;\n  font-size: 1.2rem;\n  font-weight: 300;\n  border: 1px solid ", ";\n  display:block;\n  margin-top:20px;\n\n  &:hover {\n    background-color: ", ";\n    color: ", ";\n    transform: scale(0.99);\n  }\n"])), function (_ref) {
+  var bg = _ref.bg;
+  return bg || '#fff';
+}, function (_ref2) {
+  var color = _ref2.color;
+  return color || '#333';
+}, function (_ref3) {
+  var borderColor = _ref3.borderColor;
+  return borderColor || '#303031';
+}, function (_ref4) {
+  var bgHover = _ref4.bgHover;
+  return bgHover || '#58111A';
+}, function (_ref5) {
+  var colorHover = _ref5.colorHover;
+  return colorHover || '#FFF';
+});
+var ButtonBackUp = styled_components__WEBPACK_IMPORTED_MODULE_0__["default"].div(_templateObject2 || (_templateObject2 = _taggedTemplateLiteral(["\n  border-radius: 2px;\n  cursor: pointer;\n  height: 40px;\n  width: 40px;\n  background: #58111A;\n  box-shadow: rgba(0, 0, 0, 0.16) 0px 3px 6px;\n  color: #fff;\n  display:block;\n  margin-top:20px;\n"])));
+
+/***/ }),
+
+/***/ "./resources/js/components/styles/Input.styled.js":
+/*!********************************************************!*\
+  !*** ./resources/js/components/styles/Input.styled.js ***!
+  \********************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+__webpack_require__.r(__webpack_exports__);
+/* harmony export */ __webpack_require__.d(__webpack_exports__, {
+/* harmony export */   "Input": () => (/* binding */ Input)
+/* harmony export */ });
+/* harmony import */ var styled_components__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! styled-components */ "./node_modules/styled-components/dist/styled-components.browser.esm.js");
+var _templateObject;
+
+function _taggedTemplateLiteral(strings, raw) { if (!raw) { raw = strings.slice(0); } return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }
+
+
+var Input = styled_components__WEBPACK_IMPORTED_MODULE_0__["default"].input(_templateObject || (_templateObject = _taggedTemplateLiteral(["\n    width: 100%;\n    max-width: 350px;\n    height: 40px;\n    background-color: #D8D8D8;\n    border-radius: 3px;\n    border: 1px solid #303031;\n    margin-top:20px;\n    font-family: 'Gotham Thin';\n    font-weight: 500;\n    font-style: italic;\n    padding-left:10px;\n    box-sizing: border-box;\n    font-size: 0.833rem;\n "])));
+
+/***/ }),
+
+/***/ "./resources/js/pages/TestStyle.js":
+/*!*****************************************!*\
+  !*** ./resources/js/pages/TestStyle.js ***!
+  \*****************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+__webpack_require__.r(__webpack_exports__);
+/* harmony export */ __webpack_require__.d(__webpack_exports__, {
+/* harmony export */   "default": () => (__WEBPACK_DEFAULT_EXPORT__)
+/* harmony export */ });
+/* harmony import */ var _components_styles_Button_styled__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../components/styles/Button.styled */ "./resources/js/components/styles/Button.styled.js");
+/* harmony import */ var _components_styles_Input_styled__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../components/styles/Input.styled */ "./resources/js/components/styles/Input.styled.js");
+/* harmony import */ var react_jsx_runtime__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! react/jsx-runtime */ "./node_modules/react/jsx-runtime.js");
+
+
+
+
+
+var TestStyle = function TestStyle() {
+  var test = function test() {
+    console.log('test');
+  };
+
+  return /*#__PURE__*/(0,react_jsx_runtime__WEBPACK_IMPORTED_MODULE_2__.jsxs)("div", {
+    children: [/*#__PURE__*/(0,react_jsx_runtime__WEBPACK_IMPORTED_MODULE_2__.jsx)(_components_styles_Button_styled__WEBPACK_IMPORTED_MODULE_0__.Button, {
+      onClick: function onClick() {
+        return test();
+      },
+      bg: "#fff",
+      color: "#303031",
+      colorHover: "#fff",
+      bgHover: "#303031",
+      children: "AJOUTER"
+    }), /*#__PURE__*/(0,react_jsx_runtime__WEBPACK_IMPORTED_MODULE_2__.jsx)(_components_styles_Button_styled__WEBPACK_IMPORTED_MODULE_0__.Button, {
+      bg: "#303031",
+      color: "#fff",
+      children: "BOIRE"
+    }), /*#__PURE__*/(0,react_jsx_runtime__WEBPACK_IMPORTED_MODULE_2__.jsx)(_components_styles_Input_styled__WEBPACK_IMPORTED_MODULE_1__.Input, {
+      placeholder: "Nom Usager"
+    }), /*#__PURE__*/(0,react_jsx_runtime__WEBPACK_IMPORTED_MODULE_2__.jsx)(_components_styles_Button_styled__WEBPACK_IMPORTED_MODULE_0__.ButtonBackUp, {})]
+  });
+};
+
+/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (TestStyle);
+
+/***/ })
+
+}]);
