+++ resources/js/components/styles/Global.js
@@ -1,0 +1,16 @@
+import { createGlobalStyle } from 'styled-components'
+
+//font-family: "LincolnMillerBanner-Roman"
+
+const GlobalStyles = createGlobalStyle`
+@import url('http://fonts.cdnfonts.com/css/gotham');
+@import url(//db.onlinewebfonts.com/c/a6e1f9c1f231a9d282a89b950b66509d?family=LincolnMillerBanner-Roman);
+
+  * {
+    margin:0;
+    padding:0;
+  }
+
+`
+
+export default GlobalStyles
