+++ resources/js/components/styles/Input.styled.js
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+
+
+export const Input = styled.input`
+    width: 100%;
+    max-width: 350px;
+    height: 40px;
+    background-color: #D8D8D8;
+    border-radius: 3px;
+    border: 1px solid #303031;
+    margin-top:20px;
+    font-family: 'Gotham Thin';
+    font-weight: 500;
+    font-style: italic;
+    padding-left:10px;
+    box-sizing: border-box;
+    font-size: 0.833rem;
+ `
